% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bsi_hierarchy.R
\name{extract_bsi_tables}
\alias{extract_bsi_tables}
\title{Extract data tables from Bank of Slovenia website}
\usage{
extract_bsi_tables(base_url = "https://px.bsi.si/pxweb/sl/serije_slo/")
}
\arguments{
\item{base_url}{The base URL of the Bank of Slovenia data portal}
}
\value{
A data frame containing information about top-level categories and subcategories
}
\description{
This function retrieves the top-level structure of data tables
from the Bank of Slovenia statistical data portal.
}
\examples{
\dontrun{
top_level <- extract_bsi_tables()
}
}
