% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bsi_hierarchy.R
\name{extract_second_level_tables}
\alias{extract_second_level_tables}
\title{Extract tables from second level pages}
\usage{
extract_second_level_tables(url)
}
\arguments{
\item{url}{The URL of the second-level page to extract tables from}
}
\value{
A data frame containing information about tables on the page
}
\description{
This function retrieves the detailed table information from second-level pages
of the Bank of Slovenia statistical data portal.
}
\examples{
\dontrun{
tables <- extract_second_level_tables("https://px.bsi.si/pxweb/sl/serije_slo/serije_slo__40_financni_racuni/")
}
}
