% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_px.R
\name{fetch_px}
\alias{fetch_px}
\title{Read Bank of Slovenia PX files}
\usage{
fetch_px(url, quiet = FALSE)
}
\arguments{
\item{url}{URL to PX file}

\item{quiet}{Logical. If TRUE, suppresses warnings}
}
\value{
Parsed PX object
}
\description{
Read Bank of Slovenia PX files
}
