% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_px.R
\name{get_px_data}
\alias{get_px_data}
\title{Get data values from a PX file}
\usage{
get_px_data(px_code)
}
\arguments{
\item{px_code}{Character string identifying the PX file (e.g., "F2_A1S")}
}
\value{
A vector of data values
}
\description{
Extracts the actual data values from a Bank of Slovenia PX file.
}
