% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_px.R
\name{get_px_dim_levels}
\alias{get_px_dim_levels}
\title{Get dimension levels from a PX file}
\usage{
get_px_dim_levels(px_code, con, schema = "platform")
}
\arguments{
\item{px_code}{Character string identifying the PX file (e.g., "F2_A1S")}

\item{con}{Database connection object}

\item{schema}{Database schema name, defaults to "platform"}
}
\value{
A list of dimension levels
}
\description{
Extracts the dimension levels (categories) from a Bank of Slovenia PX file.
}
