% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_px.R
\name{get_px_list}
\alias{get_px_list}
\title{Get PX file information}
\usage{
get_px_list(px_code)
}
\arguments{
\item{px_code}{Character string identifying the PX file (e.g., "F2_A1S")}
}
\value{
A PX file object as returned by pxR::read.px
}
\description{
Retrieves a PX file from the Bank of Slovenia data portal
based on the PX code.
}
