% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_category_table}
\alias{prepare_category_table}
\title{Prepare table to insert into `category` table}
\usage{
prepare_category_table(px_code, con, schema = "platform")
}
\arguments{
\item{px_code}{the original BS code (e.g. i_36_6as)}

\item{con}{a connection to the database}

\item{schema}{the schema to use for the connection, default is "platform"}
}
\value{
a dataframe with the `id`, `name`, `source_id` for each category that
the table is a member of.
}
\description{
Helper function that extracts all the parent categories from the full
hierarchy data.frame, and prepares the category table with field ids and
their names. Returns table ready to insert into the `category` table with the db_writing family
of functions. Uses the full field hierarchy with parent_ids et al, output from
\link[BS fetchR]{get_all_bsi_tables}.
}
