% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_series_table}
\alias{prepare_series_table}
\title{Prepare table to insert into `series` table}
\usage{
prepare_series_table(px_code, con, schema = "platform")
}
\arguments{
\item{px_code}{character object of the matrix code (e.g. 2300123S)}

\item{con}{connection to the database}

\item{schema}{database schema, defaults to "platform"}
}
\value{
a dataframe with the following columns: `series_title`, `series_code`,
`unit_id`, `table_id` and `interval_id`for each series in the table
well as the same number of rows as there are series
}
\description{
This is a big one. Prepares the table to insert into the series table, which
along expanding the levels to get all the series and their codes as well, which
include also figuring out their time interval, this function also tries to
extract the unit for each series,
Returns table ready to insert into the `series`table with the
db_writing family of functions.
}
