% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_unit_table}
\alias{prepare_unit_table}
\title{Prepare table to insert into `units` table}
\usage{
prepare_unit_table(px_code)
}
\arguments{
\item{px_code}{the original BS table code (e.g. i_36_6as)}
}
\value{
a dataframe with the single column containing the different units used
in this table.
}
\description{
Helper function that extracts the units for each table from the px metadata.
Returns table ready to insert into the `units`table with the
db_writing family of functions.
}
