% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_vintage_table}
\alias{prepare_vintage_table}
\title{Prepare table to insert into `vintage` table}
\usage{
prepare_vintage_table(px_code, con, schema)
}
\arguments{
\item{px_code}{the original BS code (e.g. i_36_6as)}

\item{con}{a connection to the database}

\item{schema}{the schema to use for the connection, default is "platform"}
}
\value{
a dataframe with the `series_id` and `published` columns
for all the series in this table.
}
\description{
Helper function that populates the vintage table with the new vintages. It gets
the series id's from the database and adds the publication date from the px.
}
\details{
Returns table ready to insert into the `vintage`table with the
UMARimportr::insert family of functions.
}
