% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_px.R
\name{fetch_px}
\alias{fetch_px}
\title{Read Bank of Slovenia PX files with proper encoding handling}
\usage{
fetch_px(url)
}
\arguments{
\item{url}{URL to PX file}
}
\value{
Parsed PX object or NULL on failure
}
\description{
Read Bank of Slovenia PX files with proper encoding handling
}
